fix(cli): point imports at renamed pending-rendered-height module
The previous rename commit landed the file rename but not the importer edits
(a stale pathspec aborted the git add), leaving MarkdownDisplay, useGeminiStream
and the test importing the old ./pendingRenderedHeight.js path — a module-not-
found in CI. Update the imports to the kebab-case path.
